Rewrite cache_stats from Perl to Golang
